# packages/contracts/openapi/wedge_openapi.yaml
openapi: 3.0.3
info:
  title: Wedge API
  version: 1.0.0-draft
  description: Wedge public REST API (/api), internal callback API (/internal), and canonical transport schemas. This draft
    includes report shares, checkpoint callback, and evidence packet retrieval.
servers:
- url: http://localhost:8080
  description: local API server for Swagger MVP testing
- url: https://api.wedge.example.com
  description: production placeholder
- url: https://staging-api.wedge.example.com
  description: staging placeholder
security:
- HumanBearer: []
...
paths:
...
  /api/runs/{runId}/report:
    get:
      tags:
      - Reports
      operationId: getRunReportProjection
      summary: 실행 리포트 조회
      description: Run 기준으로 생성된 리포트 projection을 조회한다. 분석 완료 후 리포트 화면이 표시할 데이터를 확인한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Run report projection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleRunReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
      - Reports
      operationId: generateRunReportProjection
      summary: 실행 리포트 생성
      description: 완료된 분석 결과를 바탕으로 리포트 projection을 생성한다. requestRunAnalysis와 Analyzer callback 이후 호출한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      responses:
        '201':
          description: Run report generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleRunReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/runs/{runId}/reports:
    get:
      tags:
      - Reports
      operationId: listRunReports
      summary: 실행 리포트 목록 조회
      description: Run에 생성된 리포트 요약 목록을 조회한다. 리포트가 여러 개 생성되는 흐름을 확인할 때 사용한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListReportSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
      - Reports
      operationId: createRunReportExport
      summary: 리포트 Markdown export 생성 또는 재사용
      description: 완료된 리포트를 즉시 Markdown artifact로 생성하거나 기존 Markdown artifact를 재사용한다. MVP에서는 format=MARKDOWN만 지원하며, 생성된 downloadUrl은 같은 origin의 artifact content URL이다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '200':
          description: Report Markdown export ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleReportExportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/reports/{reportId}:
    get:
      tags:
      - Reports
      operationId: getReport
      summary: 리포트 상세 조회
      description: reportId 기준으로 findings, nudges, decisionMap 등 상세 리포트 데이터를 확인한다. listRunReports 또는 getRunReport 응답의 id로
        테스트한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/reports/{reportId}/shares:
    get:
      tags:
      - Reports
      operationId: listReportShares
      summary: 리포트 공유 링크 목록 조회
      description: reportId에 연결된 공유 링크 목록을 조회하는 계약 API다. 현재 MVP 핵심 테스트에서는 선택 사항이다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report shares
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListReportShareResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
      - Reports
      operationId: createReportShare
      summary: 리포트 공유 링크 생성
      description: 외부 공유용 링크를 생성하거나, 만료 전 유효한 공유 링크가 이미 있으면 기존 링크를 반환한다. 만료 시간은 클라이언트가 지정하지 않으며 새로 생성되는 링크는 서버 기본값으로 생성 후 10분을 사용한다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Existing active report share reused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleReportShareResponse'
        '201':
          description: New report share created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleReportShareResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/reports/{reportId}/shares/{shareId}:
    delete:
      tags:
      - Reports
      operationId: revokeReportShare
      summary: 리포트 공유 링크 폐기
      description: reportId에 연결된 공유 링크를 폐기한다. 폐기된 token은 공개 조회에 사용할 수 없다.
      security:
      - HumanBearer: []
      parameters:
      - $ref: '#/components/parameters/ReportId'
      - name: shareId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Report share revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/report-shares/{shareToken}:
    get:
      tags:
      - Reports
      operationId: getSharedReport
      summary: 공유 token 리포트 조회
      description: 인증 헤더 없이 공유 token만으로 리포트 상세를 조회한다. share token 자체가 공개 조회 권한을 나타내는 bearer secret이며, MVP 기본 만료 시간은 생성 후 10분이다.
      security: []
      parameters:
      - name: shareToken
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Shared report detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleReportResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/report-shares/{shareToken}/artifacts/{artifactId}/content:
    get:
      tags:
      - Reports
      operationId: getSharedReportArtifactContent
      summary: 공유 token 리포트 이미지 artifact content 조회
      description: 인증 헤더 없이 공유 token만으로 공유 리포트에 연결된 run의 이미지 artifact content를 조회한다. share token이 만료/폐기되었거나 artifact가 공유 리포트의 run에 속하지 않거나 이미지 artifact가 아니면 조회할 수 없다.
      security: []
      parameters:
      - name: shareToken
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          description: Shared report artifact content
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.wedge.example.com/oauth2/authorize
          tokenUrl: https://auth.wedge.example.com/oauth2/token
          scopes:
            wedge.read: Read Wedge resources
            wedge.execute: Create and execute runs
            wedge.export: Export reports
            wedge.admin: Administrative actions
        clientCredentials:
          tokenUrl: https://auth.wedge.example.com/oauth2/token
          scopes:
            wedge.read: Read Wedge resources
            wedge.execute: Create and execute runs
            wedge.export: Export reports
            wedge.admin: Administrative actions
    InternalBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    HumanBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    DecisionStage:
      type: string
      enum:
      - FIRST_VIEW
      - VALUE
      - CTA
      - INPUT
      - COMMIT
    StageStatus:
      type: string
      enum:
      - OBSERVED
      - PASS
      - WARNING
      - BLOCKED
      - NOT_OBSERVED
      - NOT_APPLICABLE
    DecisionMapItem:
      type: object
      properties:
        stage:
          $ref: '#/components/schemas/DecisionStage'
        displayName:
          type: string
          example: 행동 선택
        status:
          $ref: '#/components/schemas/StageStatus'
        issueIds:
          type: array
          items:
            type: string
        summary:
          type: string
          nullable: true
        evidenceRefs:
          type: array
          items:
            type: string
      required:
      - stage
      - displayName
      - status
      - issueIds
      - evidenceRefs
...
    Meta:
      type: object
      properties:
        requestId:
          type: string
        correlationId:
          type: string
          nullable: true
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
          nullable: true
      required:
      - requestId
    ErrorCode:
      type: string
      enum:
      - UNAUTHORIZED
      - FORBIDDEN
      - NOT_FOUND
      - CONFLICT
      - VALIDATION_ERROR
      - DISCOVERY_NOT_FOUND
      - DISCOVERY_EXPIRED
      - DISCOVERY_FAILED
      - SCENARIO_NOT_APPLICABLE
      - SCENARIO_FIT_LOW_CONFIDENCE
      - NO_ENTRYPOINT_FOUND
      - UNSAFE_SCENARIO_ACTION
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
          - code
          - message
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - error
      - meta
...
    AckResponse:
      type: object
      properties:
        data:
          type: object
          nullable: true
          additionalProperties: true
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
...
    Artifact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        stepId:
          type: string
          format: uuid
          nullable: true
        stepKey:
          type: string
          nullable: true
        artifactType:
          type: string
          enum:
          - FRAME
          - SCREENSHOT
          - DOM_SNAPSHOT
          - AX_TREE
          - TRACE
          - HAR
          - CONSOLE_LOG
          - REPORT_PDF
          - REPORT_MARKDOWN
          - REPORT_HTML
          - REPORT_JSON
          - OTHER
        bucket:
          type: string
        key:
          type: string
        mimeType:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        sizeBytes:
          type: integer
        sha256:
          type: string
          nullable: true
        url:
          type: string
          format: uri
          nullable: true
        contentUrl:
          type: string
          description: API-relative URL for retrieving artifact bytes through Wedge. Useful for local prototype storage where
            no public object URL exists.
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - runId
      - artifactType
      - bucket
      - key
      - mimeType
      - sizeBytes
      - createdAt
...
    RunReport:
      type: object
      properties:
        runId:
          type: string
          format: uuid
        reportStatus:
          type: string
          enum:
          - NOT_READY
          - GENERATABLE
          - READY
          - FAILED
        analysisStatus:
          type: string
        analysisJobId:
          type: string
          format: uuid
          nullable: true
        reportId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          nullable: true
        format:
          type: string
          enum:
          - PDF
          - MARKDOWN
          - HTML
          - JSON
          nullable: true
        status:
          type: string
          enum:
          - GENERATING
          - READY
          - FAILED
          - ARCHIVED
          nullable: true
        summary:
          type: object
          additionalProperties: true
        decisionMap:
          type: array
          items:
            type: object
            additionalProperties: true
        findings:
          type: array
          items:
            type: object
            additionalProperties: true
        nudges:
          type: array
          items:
            type: object
            additionalProperties: true
        errorCode:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
      - runId
      - reportStatus
      - analysisStatus
      - summary
      - decisionMap
      - findings
      - nudges
...
    ReportSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        analysisJobId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        format:
          type: string
          enum:
          - PDF
          - MARKDOWN
          - HTML
          - JSON
        status:
          type: string
          enum:
          - GENERATING
          - READY
          - FAILED
          - ARCHIVED
        frictionScore:
          type: number
          nullable: true
          description: Preview score extracted from report.summary_jsonb.friction_score.
        summary:
          type: object
          additionalProperties: true
          description: Lightweight report summary projection for list and preview cards.
          nullable: true
        decisionMap:
          type: array
          items:
            $ref: '#/components/schemas/DecisionMapItem'
        topFindings:
          type: array
          description: Top 3 high-priority findings for report summary preview.
          items:
            $ref: '#/components/schemas/ReportTopFinding'
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - runId
      - title
      - format
      - status
      - decisionMap
      - topFindings
      - createdAt
    ReportTopFinding:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rank:
          type: integer
        title:
          type: string
        summary:
          type: string
        stage:
          $ref: '#/components/schemas/DecisionStage'
        severity:
          type: integer
          nullable: true
        confidence:
          type: number
          nullable: true
        priorityScore:
          type: number
          nullable: true
        previewImage:
          type: object
          allOf:
          - $ref: '#/components/schemas/ReportPreviewImage'
          nullable: true
      required:
      - id
      - rank
      - title
      - summary
      - stage
    ReportDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        analysisJobId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        format:
          type: string
          enum:
          - PDF
          - MARKDOWN
          - HTML
          - JSON
        status:
          type: string
          enum:
          - GENERATING
          - READY
          - FAILED
          - ARCHIVED
        frictionScore:
          type: number
          nullable: true
        summary:
          type: object
          additionalProperties: true
          nullable: true
        decisionMap:
          type: array
          items:
            $ref: '#/components/schemas/DecisionMapItem'
        initialDisplayCount:
          type: integer
          description: Number of high-priority findings the UI should show before the user expands more.
          example: 3
        findings:
          type: array
          items:
            $ref: '#/components/schemas/ReportDetailFinding'
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - runId
      - title
      - format
      - status
      - decisionMap
      - initialDisplayCount
      - findings
      - createdAt
    ReportDetailFinding:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rank:
          type: integer
        title:
          type: string
        summary:
          type: string
        category:
          type: string
          nullable: true
        stage:
          type: string
          nullable: true
          description: Raw finding stage. Known values follow DecisionStage, but Detail preserves stored findings.
        axis:
          type: string
          nullable: true
        severity:
          type: integer
          nullable: true
        confidence:
          type: number
          nullable: true
        priorityScore:
          type: number
          nullable: true
        impactHypothesis:
          type: string
          nullable: true
        evidenceRefs:
          type: array
          items:
            type: object
            additionalProperties: true
        previewImage:
          type: object
          allOf:
          - $ref: '#/components/schemas/ReportPreviewImage'
          nullable: true
        highlight:
          type: object
          allOf:
          - $ref: '#/components/schemas/ReportFindingHighlight'
          nullable: true
        nudges:
          type: array
          items:
            $ref: '#/components/schemas/ReportDetailNudge'
      required:
      - id
      - rank
      - title
      - summary
      - stage
      - evidenceRefs
      - nudges
    ReportFindingHighlight:
      type: object
      description: Screen coordinate metadata for drawing a finding marker over the preview image.
      properties:
        evidenceRef:
          type: string
          nullable: true
        label:
          type: string
        source:
          type: string
          enum:
          - artifact-coordinate
        coordinateSpace:
          type: string
          nullable: true
          enum:
          - viewport
          - screenshot
          - viewport_ratio
          description: Coordinate origin emitted by analyzer. `viewport` uses runner viewport coordinates.
        bounds:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            width:
              type: number
            height:
              type: number
            unit:
              type: string
              enum:
              - css_px
              - screenshot_px
              - viewport_ratio
          description: Element bounds in the unit emitted by the runner, usually css_px.
          example:
            x: 520
            y: 360
            width: 220
            height: 56
            unit: css_px
          required:
          - x
          - y
          - width
          - height
        viewport:
          type: object
          nullable: true
          properties:
            width:
              type: number
            height:
              type: number
          required:
          - width
          - height
          description: Viewport dimensions used to translate bounds into preview percentages.
          example:
            width: 1440
            height: 900
        screenshotArtifactId:
          type: string
          description: Screenshot artifact id that the coordinate was captured against.
      required:
      - label
      - source
      - bounds
      - screenshotArtifactId
    ReportDetailNudge:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rank:
          type: integer
          nullable: true
        title:
          type: string
        rationale:
          type: string
          nullable: true
        recommendation:
          type: string
          nullable: true
        difficulty:
          type: string
          nullable: true
        expectedEffect:
          type: string
          nullable: true
        validationQuestion:
          type: string
          nullable: true
      required:
      - id
      - title
    ReportPreviewImage:
      type: object
      properties:
        artifact:
          $ref: '#/components/schemas/Artifact'
        source:
          type: string
          enum:
          - STAGE_SCREENSHOT
          - REPORT_ARTIFACT
          - LATEST_SCREENSHOT
      required:
      - artifact
      - source
    ReportExport:
      type: object
      properties:
        reportId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        analysisJobId:
          type: string
          format: uuid
          nullable: true
        format:
          type: string
          enum:
          - MARKDOWN
        status:
          type: string
          enum:
          - READY
        artifactId:
          type: string
          format: uuid
        downloadUrl:
          type: string
          description: Same-origin artifact content URL.
        createdAt:
          type: string
          format: date-time
          nullable: true
      required:
      - reportId
      - runId
      - format
      - status
      - artifactId
      - downloadUrl
    ReportCreateRequest:
      type: object
      properties:
        format:
          type: string
          enum:
          - PDF
          - MARKDOWN
          - HTML
          - JSON
        analysisJobId:
          type: string
          format: uuid
          nullable: true
      required:
      - format
...
    SingleRunReportResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RunReport'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
    SingleReportResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReportDetail'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
    SingleReportExportResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReportExport'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
...
    ListReportSummaryResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReportSummary'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
...
    ReportShare:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reportId:
          type: string
          format: uuid
        shareUrl:
          type: string
          format: uri
        accessLevel:
          type: string
          enum:
          - VIEW
        expiresAt:
          type: string
          format: date-time
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - reportId
      - shareUrl
      - accessLevel
      - createdAt
    SingleReportShareResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReportShare'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
    ListReportShareResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReportShare'
        meta:
          $ref: '#/components/schemas/Meta'
      required:
      - data
      - meta
...